restructure: exceptions
