+++ microdata_validator/components/unit_types.py
@@ -1,5 +1,7 @@
 from typing import Union
 
+from microdata_validator.exceptions import UnregisteredUnitTypeError
+
 # When updating this dictionary remember to also
 # update the DatasetMetadataSchema with the
 # same key value in the enum for unitTypeType
@@ -21,7 +23,3 @@ def get_unit_id_type_for_unit_type(unit_type: str) -> Union[str, None]:
         return UNIT_ID_TYPE_FOR_UNIT_TYPE[unit_type]
     except KeyError as e:
         raise UnregisteredUnitTypeError(f'No such unit type: {str(e)}') from e
-
-
-class UnregisteredUnitTypeError(Exception):
-    pass
